refactor(backstage-plugin): Refactor data service
Reuse request logic in the data service.
Rename data service from GroupDataService to DoraDataService.
Reorganise tests to describe them better.
Rename test to describe what is being tested.

Addresses #71
