+++ backstage-plugin/plugins/open-dora/src/hooks/MetricBenchmarkHook.ts
@@ -1,18 +1,18 @@
 import { useApi } from '@backstage/core-plugin-api';
 import { useEffect, useState } from 'react';
 import { dfBenchmarkKey } from '../models/DfBenchmarkData';
-import { groupDataServiceApiRef } from '../services/GroupDataService';
+import { doraDataServiceApiRef } from '../services/DoraDataService';
 
 export const useMetricBenchmark = (type: string) => {
-  const groupDataService = useApi(groupDataServiceApiRef);
+  const doraDataService = useApi(doraDataServiceApiRef);
   const [benchmark, setDfBenchmark] = useState<dfBenchmarkKey | undefined>();
   const [error, setError] = useState<Error | undefined>();
 
   useEffect(() => {
-    groupDataService.retrieveBenchmarkData({ type: type }).then(response => {
+    doraDataService.retrieveBenchmarkData({ type: type }).then(response => {
       setDfBenchmark(response.key);
     }, setError);
-  }, [groupDataService, type]);
+  }, [doraDataService, type]);
 
   return { error: error, benchmark: benchmark };
 };
