+++ backstage-plugin/plugins/open-dora/src/services/DoraDataService.ts
@@ -1,24 +1,19 @@
 import { ConfigApi, createApiRef } from '@backstage/core-plugin-api';
-import { MetricData } from '../models/MetricData';
 import { dfBenchmarkData } from '../models/DfBenchmarkData';
+import { MetricData } from '../models/MetricData';
 
-export const groupDataServiceApiRef = createApiRef<GroupDataService>({
-  id: 'plugin.open-dora.group-data',
+export const doraDataServiceApiRef = createApiRef<DoraDataService>({
+  id: 'plugin.open-dora.data-service',
 });
 
-export class GroupDataService {
+export class DoraDataService {
   constructor(private options: { configApi: ConfigApi }) {}
 
-  async retrieveMetricDataPoints(params: {
-    type: string;
-    team?: string;
-    project?: string;
-    aggregation?: string;
-  }) {
+  private async retrieveData(params: Record<string, string>, path: string) {
     const baseUrl = this.options.configApi.getString('open-dora.apiBaseUrl');
 
     const url = new URL(baseUrl);
-    url.pathname = 'dora/api/metric';
+    url.pathname = path;
     for (const [key, value] of Object.entries(params)) {
       if (value) {
         url.searchParams.append(key, value);
@@ -32,7 +27,16 @@ export class GroupDataService {
       throw new Error(data.statusText);
     }
 
-    const response = await data.json();
+    return await data.json();
+  }
+
+  async retrieveMetricDataPoints(params: {
+    type: string;
+    team?: string;
+    project?: string;
+    aggregation?: string;
+  }) {
+    const response = await this.retrieveData(params, 'dora/api/metric');
     if (
       response.aggregation === undefined ||
       response.dataPoints === undefined
@@ -44,24 +48,8 @@ export class GroupDataService {
   }
 
   async retrieveBenchmarkData(params: { type: string }) {
-    const baseUrl = this.options.configApi.getString('open-dora.apiBaseUrl');
+    const response = await this.retrieveData(params, 'dora/api/benchmark');
 
-    const url = new URL(baseUrl);
-    url.pathname = 'dora/api/benchmark';
-    for (const [key, value] of Object.entries(params)) {
-      if (value) {
-        url.searchParams.append(key, value);
-      }
-    }
-    const data = await fetch(url.toString(), {
-      method: 'GET',
-    });
-
-    if (!data.ok) {
-      throw new Error(data.statusText);
-    }
-
-    const response = await data.json();
     if (response.key === undefined) {
       throw new Error('Unexpected response');
     }
